Rename timing helper file
Match file name convention to other helpers.
