+++ src/helpers/timing_helper.js
